+++ src/main/java/duke/Parser.java
@@ -1,3 +1,5 @@
+import tasks.TaskList;
+
 public class Parser {
     enum modifyStatus {
         MARK,
