+++ src/main/java/duke/Storage.java
@@ -1,3 +1,6 @@
+import tasks.Task;
+import tasks.TaskList;
+
 import java.io.*;
 import java.util.ArrayList;
