+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.io.Serializable;
 
 public class Task implements Serializable {
