+++ src/main/java/duke/Ui.java
@@ -1,3 +1,6 @@
+import tasks.Task;
+import tasks.TaskList;
+
 public class Ui {
     private static final String BAR = "____________________________________________________________";
